moves AuthenticatedApp up a directory for simplicity
