+++ src/sections/AuthenticatedApp.tsx
